requirements.txt
aioconsole==0.8.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.16.0
frozenlist==1.4.1
greenlet==3.0.3
identify==2.6.0
idna==3.8
multidict==6.1.0
nodeenv==1.9.1
openpyxl==3.1.5
platformdirs==4.3.2
pre-commit==3.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.4
yarl==1.11.1
